# dictionary-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dictionary API",
        "description": "Documentação da API de Dicionário",
        "contact": {
            "email": "support@coodesh.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/entries/en/{word}/favorite": {
            "post": {
                "tags": [
                    "Word"
                ],
                "summary": "Adicionar palavra aos favoritos",
                "description": "Adiciona uma palavra à lista de favoritos de um usuário.",
                "operationId": "2ef90eaa5ec698c3e20b7ba585c0b554",
                "parameters": [
                    {
                        "name": "word",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "example"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Palavra adicionada aos favoritos.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Palavra adicionada aos favoritos."
                                        },
                                        "word": {
                                            "type": "string",
                                            "example": "example"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "A palavra já está nos favoritos.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Essa palavra já está nos favoritos."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/entries/en/{word}/unfavorite": {
            "delete": {
                "tags": [
                    "Word"
                ],
                "summary": "Remover palavra dos favoritos",
                "description": "Remove uma palavra da lista de favoritos de um usuário.",
                "operationId": "5b7e645abd05c60085996f612e6f98cd",
                "parameters": [
                    {
                        "name": "word",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "example"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Palavra removida dos favoritos.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "A palavra foi removida dos favoritos."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Palavra não encontrada nos favoritos.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "A palavra não está nos favoritos."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/entries/en/{word}": {
            "get": {
                "tags": [
                    "Word"
                ],
                "summary": "Obter detalhes de uma palavra",
                "description": "Retorna os detalhes de uma palavra a partir de uma API externa.",
                "operationId": "26f4fcb3e86d6eb2c394a3db5bde95e4",
                "parameters": [
                    {
                        "name": "word",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "example"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da palavra obtidos com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "word": "example",
                                    "definition": "A representative form of a concept."
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Palavra não encontrada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Palavra não encontrada ou erro interno da API."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno no servidor.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ocorreu um erro durante o processo de busca da palavra."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/entries/en": {
            "get": {
                "tags": [
                    "Word"
                ],
                "summary": "Listar palavras do dicionário",
                "description": "Retorna uma lista de palavras com suporte a paginação e busca.",
                "operationId": "4b3138d3433bdbe347ecd1f15bdf1075",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Termo de busca",
                        "schema": {
                            "type": "string",
                            "example": "example"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limite de palavras por página",
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de palavras com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "example"
                                            }
                                        },
                                        "totalDocs": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "totalPages": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "hasNext": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "hasPrev": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}